(hel-405) extract taux de caf
