+++ src/frontend/ui/établissement-territorial-médico-social/bloc-budget-et-finances/ÉtablissementTerritorialBudgetEtFinancesMédicoSocialViewModel.tsx
@@ -1,246 +1,23 @@
-import { ChartData, ChartOptions, ScriptableScaleContext } from "chart.js";
-import { Context } from "chartjs-plugin-datalabels";
 import { ReactElement } from "react";
-import { Bar } from "react-chartjs-2";
 
 import { CadreBudgétaire } from "../../../../backend/métier/entities/établissement-territorial-médico-social/CadreBudgétaire";
 import { ÉtablissementTerritorialMédicoSocialBudgetEtFinances } from "../../../../backend/métier/entities/établissement-territorial-médico-social/ÉtablissementTerritorialMédicoSocialBudgetEtFinances";
 import { Wording } from "../../../configuration/wording/Wording";
 import { annéesManquantes, estCeLAnnéePassée } from "../../../utils/dateUtils";
 import {
-  couleurDelAbscisse,
   couleurDuFond,
-  couleurDuFondDeLaLigne,
   couleurDuFondHistogrammePrimaire,
   couleurDuFondHistogrammeSecondaire,
+  couleurErreur,
   couleurIdentifiant,
   couleurSecondPlanHistogrammeDeDépassement,
-  couleurErreur,
-  couleurDuSeuil,
 } from "../../commun/Graphique/couleursGraphique";
-import { CouleurHistogramme, GraphiqueViewModel, LibelléDeDonnéeGraphe, LibelléDeTickGraphe } from "../../commun/Graphique/GraphiqueViewModel";
+import { CouleurHistogramme, GraphiqueViewModel } from "../../commun/Graphique/GraphiqueViewModel";
 import { HistogrammeVertical } from "../../commun/Graphique/HistogrammeVertical";
 import { IndicateurTabulaire, IndicateurTabulaireProps } from "../../commun/IndicateurTabulaire/IndicateurTabulaire";
-import { MiseEnExergue } from "../../commun/MiseEnExergue/MiseEnExergue";
 import { StringFormater } from "../../commun/StringFormater";
-import { Transcription } from "../../commun/Transcription/Transcription";
 import { CompteDeResultatViewModel } from "./compte-de-resultat/CompteDeResultatViewModel";
-
-type TauxDeCaf = Readonly<{ année: number; valeur: number | null }>;
-
-export class TauxDeCafViewModel extends GraphiqueViewModel {
-  private readonly nombreDAnnéesParIndicateur = 3;
-  private readonly seuilDuTauxDeCaf = 2;
-  private readonly seuilMinimalDuTauxDeCaf = -21;
-  private readonly seuilMaximalDuTauxDeCaf = 21;
-
-  static fromBudgetFinanceMedicoSocial(budgetFinance: ÉtablissementTerritorialMédicoSocialBudgetEtFinances[], wording: Wording) {
-    const tauxDeCaf: TauxDeCaf[] = budgetFinance.map((budget) => ({ année: budget.année, valeur: budget.tauxDeCafNette.valeur }));
-    const dateMiseÀJourSource = budgetFinance.length > 0 ? budgetFinance[0].tauxDeCafNette?.dateMiseÀJourSource : "";
-    return new TauxDeCafViewModel(tauxDeCaf, dateMiseÀJourSource, wording);
-  }
-  constructor(private tauxDeCafParAnnée: TauxDeCaf[], private dateMiseÀJourSource: string, wording: Wording) {
-    super(wording);
-  }
-
-  public get leTauxDeCafEstIlRenseigné(): boolean {
-    const [années] = this.construisLesAnnéesEtSesTaux();
-
-    return années.length > 0;
-  }
-
-  private construisLesAnnéesEtSesTaux(): number[][] {
-    const valeurs: number[] = [];
-    const années: number[] = [];
-    this.tauxDeCafParAnnée.forEach((tauxDeCaf: TauxDeCaf) => {
-      const valeur = tauxDeCaf.valeur;
-      if (valeur !== null) {
-        années.push(tauxDeCaf.année);
-        valeurs.push(StringFormater.transformInRate(valeur));
-      }
-    });
-
-    return [valeurs, années];
-  }
-
-  public get tauxDeCaf(): ReactElement {
-    const [valeurs, années] = this.construisLesAnnéesEtSesTaux();
-    const construisLaCouleurDeLaBarre = (valeur: number, année: number | string): CouleurHistogramme => {
-      let premierPlan = couleurDuFondHistogrammeSecondaire;
-      let secondPlan = couleurDuFond;
-
-      if (estCeLAnnéePassée(année)) {
-        premierPlan = couleurDuFondHistogrammePrimaire;
-        secondPlan = couleurDuFond;
-      }
-
-      if (this.leTauxDeCafEstIlAberrant(valeur)) {
-        premierPlan = couleurErreur;
-        secondPlan = couleurSecondPlanHistogrammeDeDépassement;
-      }
-      return { premierPlan, secondPlan };
-    };
-    const libellésDesValeurs = valeurs.map(() => ({ couleur: couleurDuFond }));
-    const libellésDesTicks = années.map((année) => ({ tailleDePolice: estCeLAnnéePassée(année) ? this.policeGrasse : this.policeNormale }));
-
-    return this.afficheLHistogrammeDuTauxDeCaf(
-      valeurs,
-      années,
-      this.construisLesCouleursDeLHistogramme(valeurs, années, construisLaCouleurDeLaBarre),
-      libellésDesValeurs,
-      libellésDesTicks
-    );
-  }
-
-  private leTauxDeCafEstIlAberrant = (valeur: number): boolean => {
-    return valeur < this.seuilMinimalDuTauxDeCaf || valeur > this.seuilMaximalDuTauxDeCaf;
-  };
-
-  private afficheLHistogrammeDuTauxDeCaf(
-    valeurs: number[],
-    années: number[],
-    couleursDeLHistogramme: CouleurHistogramme[],
-    libellésDesValeurs: LibelléDeDonnéeGraphe[],
-    libellésDesTicks: LibelléDeTickGraphe[]
-  ) {
-    const minDeLHistogramme = Math.min(...valeurs) < this.seuilMinimalDuTauxDeCaf ? this.seuilMinimalDuTauxDeCaf : undefined;
-    const maxDeLHistogramme = Math.max(...valeurs) > this.seuilMaximalDuTauxDeCaf ? this.seuilMaximalDuTauxDeCaf : undefined;
-    const data: ChartData = {
-      datasets: [
-        {
-          backgroundColor: couleursDeLHistogramme.map((couleur) => couleur.premierPlan),
-          data: valeurs,
-          datalabels: { labels: { title: { color: libellésDesValeurs.map((libelléDeValeur) => libelléDeValeur.couleur) } } },
-          maxBarThickness: 60,
-          type: "bar",
-          xAxisID: "x",
-        },
-        {
-          borderColor: couleurDuSeuil,
-          data: [
-            { x: -1, y: this.seuilDuTauxDeCaf },
-            { x: 2, y: this.seuilDuTauxDeCaf },
-          ],
-          type: "line",
-          xAxisID: "xLine",
-        },
-        {
-          borderColor: couleurDelAbscisse,
-          data: [
-            { x: -1, y: 0 },
-            { x: 2, y: 0 },
-          ],
-          type: "line",
-          xAxisID: "xLine2",
-        },
-      ],
-      labels: années,
-    };
-    const listeAnnéesManquantes = annéesManquantes(années);
-
-    return (
-      <>
-        {annéesManquantes.length < this.nombreDAnnéesParIndicateur && (
-          <Bar
-            // @ts-ignore
-            data={data}
-            options={this.construisLesOptionsDeLHistogrammeDuTauxDeCaf(couleursDeLHistogramme, libellésDesTicks, maxDeLHistogramme, minDeLHistogramme)}
-          />
-        )}
-        {annéesManquantes.length > 0 && <MiseEnExergue>{`${this.wording.AUCUNE_DONNÉE_RENSEIGNÉE} ${listeAnnéesManquantes.join(", ")}`}</MiseEnExergue>}
-        <Transcription
-          disabled={annéesManquantes.length === this.nombreDAnnéesParIndicateur}
-          entêteLibellé={this.wording.ANNÉE}
-          identifiants={[this.wording.TAUX_DE_CAF]}
-          libellés={années}
-          valeurs={[StringFormater.addPercentToValues(valeurs)]}
-        />
-      </>
-    );
-  }
-
-  private construisLesOptionsDeLHistogrammeDuTauxDeCaf(
-    couleursDeLHistogramme: CouleurHistogramme[],
-    libellésDesTicks: LibelléDeTickGraphe[],
-    maxDeLHistogramme: number | undefined,
-    minDeLHistogramme: number | undefined
-  ): ChartOptions<"bar"> {
-    return {
-      animation: false,
-      plugins: {
-        datalabels: {
-          align: "end",
-          anchor: "start",
-          clamp: true,
-          font: {
-            family: "Marianne",
-            size: 16,
-            weight: 700,
-          },
-          formatter: (value: number, _context: Context): string => value.toLocaleString("fr") + " %",
-          textStrokeColor: couleursDeLHistogramme.map((couleur) => couleur.premierPlan),
-          textStrokeWidth: 2,
-        },
-        legend: { display: false },
-        tooltip: { enabled: false },
-      },
-      scales: {
-        x: {
-          grid: {
-            drawBorder: false,
-            drawOnChartArea: false,
-            drawTicks: false,
-          },
-          ticks: {
-            color: couleurDelAbscisse,
-            // @ts-ignore
-            font: { weight: libellésDesTicks.map((libellé) => libellé.tailleDePolice) },
-            padding: 10,
-          },
-        },
-        xLine: {
-          display: false,
-          max: 1,
-          min: 0,
-          type: "linear",
-        },
-        xLine2: {
-          display: false,
-          max: 1,
-          min: 0,
-          type: "linear",
-        },
-        y: {
-          grid: {
-            color: couleurDuFondDeLaLigne,
-            drawBorder: false,
-            drawOnChartArea: true,
-            drawTicks: false,
-          },
-          max: maxDeLHistogramme,
-          min: minDeLHistogramme,
-          ticks: {
-            callback: (tickValue: string | number) => {
-              return tickValue === this.seuilDuTauxDeCaf ? `${tickValue} (seuil)` : tickValue;
-            },
-            color: couleurDelAbscisse,
-            // @ts-ignore
-            font: {
-              weight: (context: ScriptableScaleContext) =>
-                context.tick && context.tick.value === this.seuilDuTauxDeCaf ? this.policeGrasse : this.policeNormale,
-            },
-            includeBounds: false,
-            stepSize: 2,
-          },
-        },
-      },
-    };
-  }
-
-  public get dateMiseÀJourTauxDeCaf(): string {
-    return StringFormater.formatDate(this.dateMiseÀJourSource);
-  }
-}
+import { TauxDeCafViewModel } from "./taux-de-caf/TauxDeCafViewModel";
 
 export class ÉtablissementTerritorialBudgetEtFinancesMédicoSocialViewModel extends GraphiqueViewModel {
   private readonly seuilMinimalDuTauxDeVétustéConstruction = 0;
